Refactor codebase  - Simplified form view  - Removed two different form view and unified them to single view  - Replace activity spawn to recycler view
