+++ ninchatsdk/src/main/AndroidManifest.xml
@@ -38,9 +38,6 @@
             android:name=".activities.NinchatPreAudienceQuestionnaireActivity"
             android:theme="@style/NinchatTheme.Questionnaire" />
 
-        <activity
-            android:name=".activities.NinchatComplexQuestionnaireActivity"
-            android:theme="@style/NinchatTheme.Questionnaire" />
     </application>
 
 </manifest>
