+++ ninchatsdk/src/main/java/com/ninchat/sdk/activities/NinchatPreAudienceQuestionnaireActivity.java
@@ -6,17 +6,26 @@ import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
+import android.util.Log;
 import android.view.View;
 
 import com.ninchat.sdk.NinchatSessionManager;
 import com.ninchat.sdk.R;
-import com.ninchat.sdk.adapters.NinchatComplexFormLikeQuestionnaireAdapter;
-import com.ninchat.sdk.adapters.NinchatSimpleFormLikeQuestionnaireAdapter;
+import com.ninchat.sdk.adapters.NinchatFormLikeQuestionnaireAdapter;
+import com.ninchat.sdk.events.RequireStateChange;
 import com.ninchat.sdk.helper.NinchatQuestionnaireItemDecoration;
+import com.ninchat.sdk.models.questionnaire2.NinchatQuestionnaire;
 
-import static com.ninchat.sdk.activities.NinchatComplexQuestionnaireActivity.PAGE_INDEX;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.addEof;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.isSimpleForm;
+import org.greenrobot.eventbus.EventBus;
+import org.greenrobot.eventbus.Subscribe;
+import org.greenrobot.eventbus.ThreadMode;
+import org.json.JSONArray;
+import org.json.JSONObject;
+
+import static com.ninchat.sdk.helper.NinchatQuestionnaire.getElements;
+import static com.ninchat.sdk.helper.NinchatQuestionnaire.getMatchingElement;
+import static com.ninchat.sdk.helper.NinchatQuestionnaire.getNextElement;
+import static com.ninchat.sdk.helper.NinchatQuestionnaire.getQuestionnaireElementByTarget;
 
 
 public final class NinchatPreAudienceQuestionnaireActivity extends NinchatBaseActivity {
@@ -25,6 +34,8 @@ public final class NinchatPreAudienceQuestionnaireActivity extends NinchatBaseAc
     protected static final String QUEUE_ID = "queueId";
     private String queueId;
     private int lastElement = 0;
+    private NinchatFormLikeQuestionnaireAdapter mPreAudienceQuestionnaireAdapter;
+    private RecyclerView mRecyclerView;
 
     @Override
     protected int getLayoutRes() {
@@ -38,43 +49,29 @@ public final class NinchatPreAudienceQuestionnaireActivity extends NinchatBaseAc
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        EventBus.getDefault().register(this);
         final Intent intent = getIntent();
         queueId = intent.getStringExtra(QUEUE_ID);
-        if (isSimpleForm(NinchatSessionManager.getInstance().getNinchatQuestionnaires().getNinchatPreAudienceQuestionnaire().getQuestionnaireList())) {
-            // add manual eof
-            addEof(NinchatSessionManager.getInstance().getNinchatQuestionnaires().getNinchatPreAudienceQuestionnaire().getQuestionnaireList());
-            final RecyclerView mRecyclerView = (RecyclerView) findViewById(R.id.questionnaire_form_rview);
-            mRecyclerView.setLayoutManager(new LinearLayoutManager(this));
-            final NinchatSimpleFormLikeQuestionnaireAdapter mPreAudienceQuestionnaireAdapter = new NinchatSimpleFormLikeQuestionnaireAdapter(
-                    NinchatSessionManager.getInstance().getNinchatQuestionnaires().getNinchatPreAudienceQuestionnaire(), new NinchatSimpleFormLikeQuestionnaireAdapter.Callback() {
-                @Override
-                public void onError(int position) {
-                    mRecyclerView.smoothScrollToPosition(position);
-                    close();
-                }
+        // get a list of items
+        mRecyclerView = (RecyclerView) findViewById(R.id.questionnaire_form_rview);
+        mRecyclerView.setLayoutManager(new LinearLayoutManager(this));
+        final NinchatQuestionnaire questionnaire = NinchatSessionManager
+                .getInstance()
+                .getNinchatQuestionnaires()
+                .getNinchatPreAudienceQuestionnaire();
 
-                @Override
-                public void onComplete() {
-                    close();
-                }
-            }
-            );
-            final int spaceInPixel = getResources().getDimensionPixelSize(R.dimen.items_margin_top);
-            mRecyclerView.addItemDecoration(new NinchatQuestionnaireItemDecoration(spaceInPixel));
-            mRecyclerView.setAdapter(mPreAudienceQuestionnaireAdapter);
-            mRecyclerView.setItemViewCacheSize(mPreAudienceQuestionnaireAdapter.getItemCount());
-        } else {
-            // get single item
-            // get a list of items
-            // feed them to recycler view and listen to on close or on error
-            startActivityForResult(
-                    NinchatComplexQuestionnaireActivity.getLaunchIntent(getApplicationContext(), lastElement),
-                    NinchatComplexQuestionnaireActivity.REQUEST_CODE);
-        }
+        final JSONObject groupQuestionnaire = getNextElement(questionnaire, lastElement);
+        final JSONArray questionnaireElements = getElements(groupQuestionnaire);
+        mPreAudienceQuestionnaireAdapter = new NinchatFormLikeQuestionnaireAdapter(new NinchatQuestionnaire(questionnaireElements));
+        final int spaceInPixel = getResources().getDimensionPixelSize(R.dimen.items_margin_top);
+        mRecyclerView.addItemDecoration(new NinchatQuestionnaireItemDecoration(spaceInPixel));
+        mRecyclerView.setAdapter(mPreAudienceQuestionnaireAdapter);
+        mRecyclerView.setItemViewCacheSize(mPreAudienceQuestionnaireAdapter.getItemCount());
     }
 
     @Override
     protected void onDestroy() {
+        EventBus.getDefault().unregister(this);
         super.onDestroy();
     }
 
@@ -85,22 +82,36 @@ public final class NinchatPreAudienceQuestionnaireActivity extends NinchatBaseAc
 
     @Override
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        if (requestCode == NinchatComplexQuestionnaireActivity.REQUEST_CODE) {
-            if (resultCode == RESULT_OK) {
-                final int nextIndex = data.getIntExtra(PAGE_INDEX, 0);
-                startActivityForResult(
-                        NinchatComplexQuestionnaireActivity.getLaunchIntent(getApplicationContext(),
-                                nextIndex),
-                        NinchatComplexQuestionnaireActivity.REQUEST_CODE);
-            } else if (resultCode == RESULT_CANCELED) {
-                finish();
-            }
-        }
         super.onActivityResult(requestCode, resultCode, data);
     }
 
     private void close() {
         setResult(RESULT_OK, new Intent().putExtra(NinchatPreAudienceQuestionnaireActivity.QUEUE_ID, queueId));
         finish();
+    }
+
+    private void showView(final NinchatQuestionnaire questionnaire){
+
+    }
+
+    @Subscribe(threadMode = ThreadMode.POSTING)
+    public void onEvent(RequireStateChange requireStateChange) {
+        final NinchatQuestionnaire questionnaire = NinchatSessionManager
+                .getInstance()
+                .getNinchatQuestionnaires()
+                .getNinchatPreAudienceQuestionnaire();
+
+        final JSONObject groupQuestionnaire = getNextElement(questionnaire, lastElement);
+        final String targetElement = getMatchingElement(questionnaire, groupQuestionnaire);
+        lastElement = getQuestionnaireElementByTarget(questionnaire, targetElement);
+        mPreAudienceQuestionnaireAdapter.updateContent(nextQuestionnaire(questionnaire, lastElement));
+        mRecyclerView.setAdapter(mPreAudienceQuestionnaireAdapter);
+        mPreAudienceQuestionnaireAdapter.notifyDataSetChanged();
+    }
+
+    private NinchatQuestionnaire nextQuestionnaire(final NinchatQuestionnaire questionnaire, final int index) {
+        final JSONObject groupQuestionnaire = getNextElement(questionnaire, index);
+        final JSONArray questionnaireElements = getElements(groupQuestionnaire);
+        return new NinchatQuestionnaire(questionnaireElements);
     }
 }
