+++ ninchatsdk/src/main/java/com/ninchat/sdk/adapters/NinchatFormLikeQuestionnaireAdapter.java
@@ -2,23 +2,24 @@ package com.ninchat.sdk.adapters;
 
 import android.support.annotation.NonNull;
 import android.support.v7.widget.RecyclerView;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.ViewGroup;
 
 import com.ninchat.sdk.R;
+import com.ninchat.sdk.adapters.holders.formview.NinchatButtonViewHolder;
 import com.ninchat.sdk.adapters.holders.formview.NinchatCheckboxViewHolder;
-import com.ninchat.sdk.adapters.holders.formview.NinchatControlFlowViewHolder;
 import com.ninchat.sdk.adapters.holders.formview.NinchatDropDownSelectViewHolder;
+import com.ninchat.sdk.adapters.holders.formview.NinchatInputFieldViewHolder;
 import com.ninchat.sdk.adapters.holders.formview.NinchatLikeRtViewHolder;
 import com.ninchat.sdk.adapters.holders.formview.NinchatRadioBtnViewHolder;
-import com.ninchat.sdk.adapters.holders.formview.NinchatInputFieldViewHolder;
 import com.ninchat.sdk.adapters.holders.formview.NinchatTextViewHolder;
 import com.ninchat.sdk.models.questionnaire2.NinchatQuestionnaire;
 
 import org.json.JSONObject;
 
+import static com.ninchat.sdk.helper.NinchatQuestionnaire.BUTTON;
 import static com.ninchat.sdk.helper.NinchatQuestionnaire.CHECKBOX;
-import static com.ninchat.sdk.helper.NinchatQuestionnaire.EOF;
 import static com.ninchat.sdk.helper.NinchatQuestionnaire.INPUT;
 import static com.ninchat.sdk.helper.NinchatQuestionnaire.LIKERT;
 import static com.ninchat.sdk.helper.NinchatQuestionnaire.RADIO;
@@ -28,16 +29,12 @@ import static com.ninchat.sdk.helper.NinchatQuestionnaire.TEXT_AREA;
 import static com.ninchat.sdk.helper.NinchatQuestionnaire.UNKNOWN;
 import static com.ninchat.sdk.helper.NinchatQuestionnaire.getItemType;
 
-public class NinchatSimpleFormLikeQuestionnaireAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
-
-    private final String TAG = NinchatSimpleFormLikeQuestionnaireAdapter.class.getSimpleName();
+public class NinchatFormLikeQuestionnaireAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
+    private final String TAG = NinchatFormLikeQuestionnaireAdapter.class.getSimpleName();
     private NinchatQuestionnaire questionnaire;
-    private Callback callback;
 
-    public NinchatSimpleFormLikeQuestionnaireAdapter(final NinchatQuestionnaire questionnaire,
-                                                     final Callback callback) {
-        this.questionnaire = questionnaire;
-        this.callback = callback;
+    public NinchatFormLikeQuestionnaireAdapter(final com.ninchat.sdk.models.questionnaire2.NinchatQuestionnaire ninchatPreAudienceQuestionnaire) {
+        this.questionnaire = ninchatPreAudienceQuestionnaire;
     }
 
     @Override
@@ -82,10 +79,10 @@ public class NinchatSimpleFormLikeQuestionnaireAdapter extends RecyclerView.Adap
                 return new NinchatLikeRtViewHolder(
                         LayoutInflater.from(parent.getContext()).inflate(R.layout.dropdown_with_label, parent, false),
                         position, questionnaire);
-            case EOF:
-                return new NinchatControlFlowViewHolder(
+            case BUTTON:
+                return new NinchatButtonViewHolder(
                         LayoutInflater.from(parent.getContext()).inflate(R.layout.control_buttons, parent, false),
-                        currentItem, controlFlowCallback);
+                        position, questionnaire);
 
         }
         return null;
@@ -93,47 +90,15 @@ public class NinchatSimpleFormLikeQuestionnaireAdapter extends RecyclerView.Adap
 
     @Override
     public void onBindViewHolder(@NonNull RecyclerView.ViewHolder viewHolder, int position) {
-        if (viewHolder instanceof NinchatTextViewHolder) {
-            ((NinchatTextViewHolder) viewHolder).bind(position, questionnaire);
-        } else if (viewHolder instanceof NinchatInputFieldViewHolder) {
-            ((NinchatInputFieldViewHolder) viewHolder).bind();
-        } else if (viewHolder instanceof NinchatRadioBtnViewHolder) {
-            ((NinchatRadioBtnViewHolder) viewHolder).update();
-        } else if (viewHolder instanceof NinchatDropDownSelectViewHolder) {
-            ((NinchatDropDownSelectViewHolder) viewHolder).bind();
-        } else if (viewHolder instanceof NinchatCheckboxViewHolder) {
-            ((NinchatCheckboxViewHolder) viewHolder).update();
-        } else if (viewHolder instanceof NinchatLikeRtViewHolder) {
-            ((NinchatLikeRtViewHolder) viewHolder).bind();
-        }
+        Log.d(TAG, "on bind view holder " + position);
+    }
+
+    public void updateContent(final NinchatQuestionnaire ninchatQuestionnaire) {
+        this.questionnaire = ninchatQuestionnaire;
     }
 
     @Override
     public int getItemCount() {
         return questionnaire.size();
-    }
-
-    private NinchatControlFlowViewHolder.Callback controlFlowCallback = new NinchatControlFlowViewHolder.Callback() {
-        @Override
-        public void onClickNext() {
-            final int errorIndex = questionnaire.updateRequiredFieldStats();
-            if (errorIndex != -1) {
-                notifyDataSetChanged();
-                callback.onError(errorIndex);
-            } else {
-                callback.onComplete();
-            }
-        }
-
-        @Override
-        public void onClickPrevious() {
-            // todo implement
-        }
-    };
-
-    public interface Callback {
-        void onError(final int position);
-
-        void onComplete();
     }
 }
